split implementation into separate files (pre jsi cleanup)
